feat(docs): prerender every route to static HTML
Replaces the SPA build plus the head-rewriting `vite-plugin-seo` with a real
SSG pass: `scripts/prerender.mjs` builds the client bundle, builds an SSR
bundle from `entry-server.tsx`, then renders each of the 81 routes to its own
`dist/<slug>/index.html` with a per-route `<head>` (`src/seo/head.ts`) and a
generated sitemap. `entry-client.tsx` hydrates whatever markup it finds.

Crawlers now get the real page instead of an empty `#root` plus a link
fallback, so the `.docs-seo-fallback` block and its `data-js` no-flash guard
are gone. `vite-plugin-doc-dates` stamps each page's last-modified date from
git for the sitemap. The Netlify catch-all stays as a last-resort fallback —
every real route is a static file now.

Also fixes two hydration hazards the prerender surfaced: `useTheme` had no
server snapshot, so it now reports the `dark` baseline baked into index.html
and re-reads the real theme right after hydrating; and prose on its own lines
inside `Card.Description` parsed as a markdown block, nesting a `<p>` inside a
`<p>` that the browser un-nests while parsing the prerendered HTML.
