+++ packages/docs/netlify.toml
@@ -26,9 +26,11 @@
   status = 301
   force = true
 
-# SPA fallback for the docs client-side routes (e.g. /components/button).
-# Netlify serves existing static files before applying redirects, so everything
-# under /storybook (real files) is unaffected by this catch-all.
+# Last-resort fallback only. Every real route is prerendered to its own
+# dist/<slug>/index.html (see packages/docs/scripts/prerender.mjs) and Netlify
+# serves existing static files before applying redirects — so this catch-all is
+# reached only by URLs that match no page, and everything under /storybook (real
+# files) is unaffected by it.
 [[redirects]]
   from = "/*"
   to = "/index.html"
